+++ CLD/Presentation /Auth/ViewController/AuthViewController.swift
@@ -10,11 +10,10 @@ import UIKit
 import SnapKit
 import RxSwift
 import RxCocoa
-import KakaoSDKUser
-import FBSDKLoginKit
 
 final class AuthViewController: BaseViewController {
     let signView = SignView()
+    private var AuthUseCase: AuthUseCase = CLD.AuthUseCase()
     
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -35,7 +34,7 @@ final class AuthViewController: BaseViewController {
     private func buttonTap() {
         signView.kakaoButton.rx.tap
             .bind {
-                self.signInKakao()
+                self.AuthUseCase.KakaoSignin()
          }.disposed(by: disposeBag)
         
         signView.appleButton.rx.tap
@@ -45,59 +44,7 @@ final class AuthViewController: BaseViewController {
         
         signView.instaButton.rx.tap
             .bind {
-                self.loginButton()
+                self.AuthUseCase.FBSignin(AuthViewController())
             }.disposed(by: disposeBag)
     }
-    
-    private func signInKakao() {
-        // isKakaoTalkLoginAvailable() : 카톡 설치 되어있으면 true
-        if (UserApi.isKakaoTalkLoginAvailable()) {
-            //카톡 설치되어있으면 -> 카톡으로 로그인
-            UserApi.shared.loginWithKakaoTalk {(oauthToken, error) in
-                if let error = error {
-                    print(error)
-                } else {
-                    print("카카오 톡으로 로그인 성공")
-                    
-                    guard let authInfo = oauthToken else { return }
-                    let sceneDelegate = UIApplication.shared.connectedScenes.first?.delegate as? SceneDelegate
-                    sceneDelegate?.changeRootView()
-                    print("== Kakao Login \(authInfo)")
-                }
-            }
-        } else {
-            // 카톡 없으면 -> 계정으로 로그인
-            UserApi.shared.loginWithKakaoAccount { (oauthToken, error) in
-                if let error = error {
-                    print(error)
-                } else {
-                    print("카카오 계정으로 로그인 성공")
-                    
-                    guard let authInfo = oauthToken else { return }
-                    let sceneDelegate = UIApplication.shared.connectedScenes.first?.delegate as? SceneDelegate
-                    sceneDelegate?.changeRootView()
-                    print("== Kakao Login \(authInfo)")
-                }
-            }
-        }
-    }
-    
-    func loginButton() {
-            let loginManager = LoginManager()
-            loginManager.logIn(permissions: ["public_profile"], from: self) { result, error in
-                if let error = error {
-                    print("Encountered Erorr: \(error)")
-                } else {
-                    if let result = result {
-                        print("페이스북 계정으로 로그인 성공")
-                        let tokenString = result.token?.tokenString
-                        let userID = result.token?.userID
-                        print("token: \(userID)")
-                        
-                        let sceneDelegate = UIApplication.shared.connectedScenes.first?.delegate as? SceneDelegate
-                        sceneDelegate?.changeRootView()
-                    }
-                }
-            }
-        }
 }
